stock-app/requirements.txt
streamlit==1.38.0
pandas>=2.1
python-dotenv>=1.0
snowflake-connector-python>=3.10
plotly>=5.22
